requirements.txt
fastapi
uvicorn
pydantic
faiss-cpu
strawberry-graphql
ollama
python-multipart
pyyaml


